Inheritance
This folder contains 1 project. The main file, CashRegister uses
MoneyConverter and Receipt. Users are prompted to either enter a
pre-existing list (.txt file) or individually input items and their
prices. After the user is finished, the program prints out a receipt.
The samples are what the program should look like when compiled.