Objects
The program is run through Restaurant. Users will be able to add or
delete items on a menu and give it a price and the calories. Then the
menu is either deleted or inserted into an array in Menu. The array will
automatically resize when the amount of items becomes too large. Users
can also print out the items according to calorie range or price as well
various other commands.