Parking Lot
This is a small text based game where users input a pre-existing map
files such as the ones provided in the lot#. From there, the user uses
"wasd" to navigate the care "U" to the bottom row. The map is stored in
arrays and the program keeps track of where the user's coordinates are
to update the map.