Movie Theater
This Project focuses on using Inheritance by having several movie
theaters in a town and many movies, of similar title as well, among the
many movie theaters. The program allows the user to input a movie that
contains the price and times, search for a movie title, print movies by
categories, and print movies at a certain time.